# perf: Optimization pass for reading from dataobjs