--- a/pkg/dataobj/internal/dataset/value_encoding_plain_test.go
+++ b/pkg/dataobj/internal/dataset/value_encoding_plain_test.go
@@ -21,12 +21,12 @@ var testStrings = []string{
 
 var batchSize = 64
 
-func Test_plainEncoder(t *testing.T) {
+func Test_plainStringEncoder(t *testing.T) {
 	var buf bytes.Buffer
 
 	var (
-		enc    = newPlainEncoder(&buf)
-		dec    = newPlainDecoder(&buf)
+		enc    = newPlainStringEncoder(&buf)
+		dec    = newPlainStringDecoder(&buf)
 		decBuf = make([]Value, batchSize)
 	)
 
@@ -51,12 +51,12 @@ func Test_plainEncoder(t *testing.T) {
 	require.Equal(t, testStrings, out)
 }
 
-func Test_plainEncoder_partialRead(t *testing.T) {
+func Test_plainStringEncoder_partialRead(t *testing.T) {
 	var buf bytes.Buffer
 
 	var (
-		enc    = newPlainEncoder(&buf)
-		dec    = newPlainDecoder(&oneByteReader{&buf})
+		enc    = newPlainStringEncoder(&buf)
+		dec    = newPlainStringDecoder(&oneByteReader{&buf})
 		decBuf = make([]Value, batchSize)
 	)
 
@@ -81,8 +81,8 @@ func Test_plainEncoder_partialRead(t *testing.T) {
 	require.Equal(t, testStrings, out)
 }
 
-func Benchmark_plainEncoder_Append(b *testing.B) {
-	enc := newPlainEncoder(streamio.Discard)
+func Benchmark_plainStringEncoder_Append(b *testing.B) {
+	enc := newPlainStringEncoder(streamio.Discard)
 
 	for i := 0; i < b.N; i++ {
 		for _, v := range testStrings {
@@ -91,17 +91,114 @@ func Benchmark_plainEncoder_Append(b *testing.B) {
 	}
 }
 
-func Benchmark_plainDecoder_Decode(b *testing.B) {
+func Benchmark_plainStringDecoder_Decode(b *testing.B) {
 	buf := bytes.NewBuffer(make([]byte, 0, 1024)) // Large enough to avoid reallocations.
 
 	var (
-		enc    = newPlainEncoder(buf)
-		dec    = newPlainDecoder(buf)
+		enc    = newPlainStringEncoder(buf)
+		dec    = newPlainStringDecoder(buf)
 		decBuf = make([]Value, batchSize)
 	)
 
 	for _, v := range testStrings {
-		require.NoError(b, enc.Encode(StringValue(v)))
+		require.NoError(b, enc.Encode(ByteArrayValue([]byte(v))))
+	}
+
+	var err error
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		for {
+			_, err = dec.Decode(decBuf[:batchSize])
+			if errors.Is(err, io.EOF) {
+				break
+			} else if err != nil {
+				b.Fatal(err)
+			}
+		}
+	}
+}
+
+func Test_plainBytesEncoder(t *testing.T) {
+	var buf bytes.Buffer
+
+	var (
+		enc    = newPlainBytesEncoder(&buf)
+		dec    = newPlainBytesDecoder(&buf)
+		decBuf = make([]Value, batchSize)
+	)
+
+	for _, v := range testStrings {
+		require.NoError(t, enc.Encode(ByteArrayValue([]byte(v))))
+	}
+
+	var out []string
+
+	for {
+		n, err := dec.Decode(decBuf[:batchSize])
+		if errors.Is(err, io.EOF) {
+			break
+		} else if err != nil {
+			t.Fatal(err)
+		}
+		for _, v := range decBuf[:n] {
+			out = append(out, string(v.ByteArray()))
+		}
+	}
+
+	require.Equal(t, testStrings, out)
+}
+
+func Test_plainBytesEncoder_partialRead(t *testing.T) {
+	var buf bytes.Buffer
+
+	var (
+		enc    = newPlainBytesEncoder(&buf)
+		dec    = newPlainBytesDecoder(&oneByteReader{&buf})
+		decBuf = make([]Value, batchSize)
+	)
+
+	for _, v := range testStrings {
+		require.NoError(t, enc.Encode(ByteArrayValue([]byte(v))))
+	}
+
+	var out []string
+
+	for {
+		n, err := dec.Decode(decBuf[:batchSize])
+		if errors.Is(err, io.EOF) {
+			break
+		} else if err != nil {
+			t.Fatal(err)
+		}
+		for _, v := range decBuf[:n] {
+			out = append(out, string(v.ByteArray()))
+		}
+	}
+
+	require.Equal(t, testStrings, out)
+}
+
+func Benchmark_plainBytesEncoder_Append(b *testing.B) {
+	enc := newPlainBytesEncoder(streamio.Discard)
+
+	for i := 0; i < b.N; i++ {
+		for _, v := range testStrings {
+			_ = enc.Encode(ByteArrayValue([]byte(v)))
+		}
+	}
+}
+
+func Benchmark_plainBytesDecoder_Decode(b *testing.B) {
+	buf := bytes.NewBuffer(make([]byte, 0, 1024)) // Large enough to avoid reallocations.
+
+	var (
+		enc    = newPlainBytesEncoder(buf)
+		dec    = newPlainBytesDecoder(buf)
+		decBuf = make([]Value, batchSize)
+	)
+
+	for _, v := range testStrings {
+		require.NoError(b, enc.Encode(ByteArrayValue([]byte(v))))
 	}
 
 	var err error