--- a/pkg/dataobj/internal/sections/logs/table_build.go
+++ b/pkg/dataobj/internal/sections/logs/table_build.go
@@ -27,7 +27,7 @@ func buildTable(buf *tableBuffer, pageSize int, compressionOpts dataset.Compress
 
 		_ = streamIDBuilder.Append(i, dataset.Int64Value(record.StreamID))
 		_ = timestampBuilder.Append(i, dataset.Int64Value(record.Timestamp.UnixNano()))
-		_ = messageBuilder.Append(i, dataset.StringValue(record.Line))
+		_ = messageBuilder.Append(i, dataset.StringValueFromBytes(record.Line))
 
 		for _, md := range record.Metadata {
 			metadataBuilder := buf.Metadata(md.Name, pageSize, compressionOpts)