**What this PR does / why we need it**:
Contains some optimizations for reading logs from a dataobj.
* Adds option to get String values as a byte slice of the underlying data instead of allocating a string
* Uses labels.Labels for structured metadata instead of push.LabelsAdaptor.
  * push.LabelsAdaptor comes from Kafka, so we pay the conversion cost once during Append rather than once per query (assuming we read more than we write)
* Use the Process over ProcessString methods in LogQL pipelines, utilising the byte slice we got.

I ran a logs reader benchmark on a dataobj containing 10,000 log lines, each with 2 structured metadata attached. We get a 26% speed increase which 90% less data allocated. It still allocates 7 times per log line. 5 of these are from structured metadata. All this happens after predicate matching takes place so it will only affect lines matching any predicates.

```
$benchstat before.txt after.txt                                                                                             
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/dataobj
cpu: Apple M3 Max
              │ before.txt  │              after.txt              │
              │   sec/op    │   sec/op     vs base                │
LogsReader-14   3.777m ± 2%   2.792m ± 1%  -26.06% (p=0.000 n=10)

              │  before.txt   │              after.txt               │
              │     B/op      │     B/op      vs base                │
LogsReader-14   23.465Mi ± 2%   2.057Mi ± 8%  -91.23% (p=0.000 n=10)

              │  before.txt  │              after.txt              │
              │  allocs/op   │  allocs/op   vs base                │
LogsReader-14   100.56k ± 0%   70.47k ± 0%  -29.92% (p=0.000 n=10)
```

There might be something we can do to intern label keys/values, but I didn't want to go too far in one go.